add validation role in use

+++ src/main/java/eoen/jwtroles/services/RoleService.java
@@ -1,5 +1,6 @@
 package eoen.jwtroles.services;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,15 +9,20 @@ import org.springframework.stereotype.Service;
 import org.springframework.data.domain.Pageable;
 
 import eoen.jwtroles.entities.Role;
+import eoen.jwtroles.entities.User;
 import eoen.jwtroles.exception.BdException;
 import eoen.jwtroles.exception.EntityNotFoundException;
 import eoen.jwtroles.repositories.RoleRepository;
+import eoen.jwtroles.repositories.UserRepository;
 
 @Service
 public class RoleService {
 
     @Autowired
     private RoleRepository roleRepository;
+
+    @Autowired
+    private UserRepository userRepository;
 
     public Role postNewRole(Role role) {
         Optional<Role> roleActive = roleRepository.findByRolename(role.getRoleName());
@@ -38,22 +44,37 @@ public class RoleService {
     }
 
     public void deleteRole(Long id) {
-        Role baseRole = this.getRoleById(id);
-        // logica para validar se a role esta em uso
-        roleRepository.delete(baseRole);
+        Role baseRole = getRoleById(id);
+        System.out.println(baseRole);
+        if (roleInUse(baseRole)) {
+            throw new BdException("Role In Use!");
+        } else {
+            roleRepository.delete(baseRole);
+        }
+    }
+
+    private Boolean roleInUse(Role baseRole) {
+        List<User> users = userRepository.findAll();
+        for (User user : users) {
+            if (user.getRole().contains(baseRole)) {
+                return true;
+            }
+        }
+        return false;
     }
 
     public Role updateRole(Role role, Long id) {
 
-        System.out.println("----------------------");
-        // logica para validar se a role esta em uso
-
         Role baseRole = this.getRoleById(id);
-        role.setRoleId(baseRole.getRoleId());
+        if (roleInUse(baseRole)) {
+            throw new BdException("Role In Use!");
+        } else {
+            role.setRoleId(baseRole.getRoleId());
 
-        Optional<Role> roleActive = roleRepository.findByRolename(role.getRoleName());
-        if (roleActive.isPresent() && baseRole.getRoleName() != roleActive.get().getRoleName())
-            throw new BdException("roleName already exists without bd!");
-        return roleRepository.save(role);
+            Optional<Role> roleActive = roleRepository.findByRolename(role.getRoleName());
+            if (roleActive.isPresent() && baseRole.getRoleName() != roleActive.get().getRoleName())
+                throw new BdException("roleName already exists without bd!");
+            return roleRepository.save(role);
+        }
     }
 }
